Edit task 2. Deploy archive

+++ 2-do_deploy_web_static.py
@@ -1,8 +1,7 @@
 #!/usr/bin/python3
 """
-Fabric script (based on the file
-1-pack_web_static.py) that distributes an
-archive to your web servers
+Fabric script that distributes
+an archive to the web servers
 """
 from fabric.api import *
 import os
@@ -23,24 +22,41 @@ def do_deploy(archive_path):
         file_name = archive_path.split("/")[-1]
         no_ext = file_name.split(".")[0]
         """Upload the archive"""
-        put(archive_path, "/tmp/")
+        tar_file = put(archive_path, "/tmp/")
+        if tar_file.failed:
+            return False
         """Uncompress the archive"""
-        run("mkdir -p /data/web_static/releases/{}/".format(no_ext))
-        run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/"
-            .format(file_name, no_ext))
+        tar_file = run("mkdir -p /data/web_static/releases/{}/".format(no_ext))
+        if tar_file.failed:
+            return False
+        tar_file = run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/"
+                       .format(file_name, no_ext))
+        if tar_file.failed:
+            return False
         """Delete the archive"""
-        run("rm /tmp/{}".format(file_name))
+        tar_file = run("rm /tmp/{}".format(file_name))
+        if tar_file.failed:
+            return False
         """Move contents to host web_static"""
-        run('mv /data/web_static/releases/{}/web_static/* \
-/data/web_static/releases/{}/'.format(no_ext, no_ext))
+        tar_file = run('mv /data/web_static/releases/{}/web_static/* \
+                /data/web_static/releases/{}/'.format(no_ext, no_ext))
+        if tar_file.failed:
+            return False
         """Remove web_static dir"""
-        run('rm -rf /data/web_static/releases/{}/web_static'
-            .format(no_ext))
+        tar_file = run('rm -rf /data/web_static/releases/{}/web_static'
+                       .format(no_ext))
+        if tar_file.failed:
+            return False
         """Delete the symbolic link"""
-        run("rm -rf /data/web_static/current")
+        tar_file = run("rm -rf /data/web_static/current")
+        if tar_file.failed:
+            return False
         """Create a new symbolic link"""
-        run("ln -s /data/web_static/releases/{} /data/web_static/current"
-            .format(no_ext))
+        tar_file = run("ln -s /data/web_static/releases/{} \
+                /data/web_static/current".format(no_ext))
+        if tar_file.failed:
+            return False
+        print('New version deployed!')
         return True
     except Exception:
         return False
